webservices/webservices-rest-com-jaxrs-e-jersey/src/main/java/br/com/alura/loja/modelo/Projeto.java
package br.com.alura.loja.modelo;

import com.google.gson.Gson;
import com.thoughtworks.xstream.XStream;

public class Projeto {

	private long id;
	private String nome;
	private int anoDeInicio;

	public Projeto(long id, String nome, int anoDeInicio) {
		this.id = id;
		this.nome = nome;
		this.anoDeInicio = anoDeInicio;
	}

	public long getId() {
		return id;
	}

	public void setId(long id) {
		this.id = id;
	}

	public Object getNome() {
		return nome;
	}

	public String toXML() {
		return new XStream().toXML(this);
	}

	public String toJson() {
		return new Gson().toJson(this);
	}

	public static Projeto fromXML(String conteudo) {
		return (Projeto) new XStream().fromXML(conteudo);
	}

}
